move KL div implementation closer to Distributions
